requirements.txt
kafka-python
matplotlib
asyncpg
python-dotenv
psycopg2-binary